% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_gov_types.R
\name{filter_gov_types}
\alias{filter_gov_types}
\title{Filter government types in country_meta_data.}
\usage{
filter_gov_types(
  types = c("Pure Democracy", "Pure Monarchy", "Democratic Monarchy", "Communist",
    "Other")
)
}
\arguments{
\item{types}{Optional character vector of government types to be selected.}
}
\value{
A tibble of the country_meta_data dataset filtered by selected government types.
}
\description{
Input a vector of gov_type levels to be selected.
}
