% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_regimes.R
\name{filter_regimes}
\alias{filter_regimes}
\title{Filter regime categories in country_meta_data.}
\usage{
filter_regimes(
  regimes = c("Parliamentary democracy", "Civilian dictatorship",
    "Presidential democracy", "Military dictatorship", "Mixed democratic",
    "Royal dictatorship")
)
}
\arguments{
\item{types}{Optional character vector of regime categories to be selected.}
}
\value{
A tibble of the country_meta_data dataset filtered by selected regime categories.
}
\description{
Input a vector of regime_category levels to be selected.
}
