% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load geopolitical dataset}
\usage{
load_data()
}
\arguments{
\item{path}{Path to the CSV file.}
}
\value{
A tibble containing the dataset.
}
\description{
Reads a CSV file into a tibble.
}
