% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_welch_anova.R
\name{log_welch_anova}
\alias{log_welch_anova}
\title{Welch's ANOVA table between a categorical variable and a log-transformed quantitative variable}
\usage{
log_welch_anova(quant_var, cat_var)
}
\arguments{
\item{quant_var}{A quantitative variable between 'GDP' and 'Population'}

\item{cat_var}{A categorical variable between 'gov_type' and 'regime_category'.}
}
\value{
A Welch's ANOVA table.
}
\description{
Creates a Welch's ANOVA table between a categorical variable and the logarithm of a quantitative variable.
}
