% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_gdp_by_category.R
\name{summarize_gdp_by_category}
\alias{summarize_gdp_by_category}
\title{Summarizes a categorical variable and log(GDP)}
\usage{
summarize_gdp_by_category(cat_var)
}
\arguments{
\item{cat_var}{A categorical variable.}
}
\value{
A summary table.
}
\description{
Creates a summary table between a categorical variable and log(GDP).
}
