% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_gdp_by_category.R
\name{summarize_gdp_by_category}
\alias{summarize_gdp_by_category}
\title{Summarizes a categorical variable and log(GDP)}
\usage{
summarize_gdp_by_category(cat_var, gov_types = NULL, regimes = NULL)
}
\arguments{
\item{cat_var}{A categorical variable between 'gov_type' and 'regime_category'.}

\item{gov_types}{Optional character vector of government types to be selected.}

\item{regimes}{Optional character vector of regime categories to be selected.}
}
\value{
A gt summary table.
}
\description{
Creates a summary table between a categorical variable and log(GDP).
}
